[sandbox][string] Add String::Utf8Length and String::EncodeUtf8
Bug: 373485796, 376071292

+++ src/diagnostics/perf-jit.cc
@@ -222,7 +222,8 @@ uint64_t LinuxPerfJitLogger::GetTimestamp() {
 
 void LinuxPerfJitLogger::LogRecordedBuffer(
     Tagged<AbstractCode> abstract_code,
-    MaybeHandle<SharedFunctionInfo> maybe_sfi, const char* name, int length) {
+    MaybeHandle<SharedFunctionInfo> maybe_sfi, const char* name,
+    size_t length) {
   DisallowGarbageCollection no_gc;
   if (v8_flags.perf_basic_prof_only_functions) {
     CodeKind code_kind = abstract_code->kind(isolate_);
@@ -264,7 +265,7 @@ void LinuxPerfJitLogger::LogRecordedBuffer(
 
 #if V8_ENABLE_WEBASSEMBLY
 void LinuxPerfJitLogger::LogRecordedBuffer(const wasm::WasmCode* code,
-                                           const char* name, int length) {
+                                           const char* name, size_t length) {
   base::LockGuard<base::RecursiveMutex> guard_file(GetFileMutex().Pointer());
 
   if (perf_output_handle_ == nullptr) return;
@@ -279,10 +280,11 @@ void LinuxPerfJitLogger::LogRecordedBuffer(const wasm::WasmCode* code,
 void LinuxPerfJitLogger::WriteJitCodeLoadEntry(const uint8_t* code_pointer,
                                                uint32_t code_size,
                                                const char* name,
-                                               int name_length) {
+                                               size_t name_length) {
   PerfJitCodeLoad code_load;
   code_load.event_ = PerfJitCodeLoad::kLoad;
-  code_load.size_ = sizeof(code_load) + name_length + 1 + code_size;
+  code_load.size_ = base::checked_cast<uint32_t>(sizeof(code_load) +
+                                                 name_length + 1 + code_size);
   code_load.time_stamp_ = GetTimestamp();
   code_load.process_id_ = static_cast<uint32_t>(process_id_);
   code_load.thread_id_ = static_cast<uint32_t>(base::OS::GetCurrentThreadId());
@@ -294,7 +296,7 @@ void LinuxPerfJitLogger::WriteJitCodeLoadEntry(const uint8_t* code_pointer,
   code_index_++;
 
   LogWriteBytes(reinterpret_cast<const char*>(&code_load), sizeof(code_load));
-  LogWriteBytes(name, name_length);
+  LogWriteBytes(name, static_cast<uint32_t>(name_length));
   LogWriteBytes(kStringTerminator, sizeof(kStringTerminator));
   LogWriteBytes(reinterpret_cast<const char*>(code_pointer), code_size);
 }
@@ -317,7 +319,7 @@ base::Vector<const char> GetScriptName(Tagged<Object> maybeScript,
       return {reinterpret_cast<char*>(str->GetChars(no_gc)),
               static_cast<size_t>(str->length())};
     } else if (IsString(name_or_url)) {
-      uint32_t length;
+      size_t length;
       *storage =
           Cast<String>(name_or_url)
               ->ToCString(DISALLOW_NULLS, FAST_STRING_TRAVERSAL, &length);
